refactor: remove only the parent folder named errors
