+++ src/domain/shared/exceptions/base.py
