+++ src/use_cases/order/find/find_order.py
@@ -3,7 +3,7 @@ from dataclasses import asdict
 from src.domain.aggregates.order.interfaces.order_repository import (
     OrderRepositoryInterface,
 )
-from src.domain.shared.errors.exceptions.order import (
+from src.domain.shared.exceptions.order import (
     OrderNotFoundException,
 )
 from src.use_cases.order.find.find_order_dto import (
